move template into product forms
