+++ src/domain/products/product-form/sections/sales-channels/product-availability-modal/available-channels-modal-screen.tsx
